+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Each release heading covers the **Project** version. Subsystem version bumps are
 - Merge branch 'master' of https://github.com/tombigel/sticky-playground
 - add Wix animation audit skills
 - add Wix library truth audit workflow
+- upgrade Wix animation libraries
 
 ### Fixed